use Svnapot in hugetlb page

+++ arch/riscv/include/asm/pgtable.h
@@ -308,7 +308,7 @@ static inline int pte_special(pte_t pte)
 	return pte_val(pte) & _PAGE_SPECIAL;
 }
 
-static inline int pte_napot(pte_t pte)
+static inline unsigned long pte_napot(pte_t pte)
 {
 	return pte_val(pte) & _PAGE_NAPOT;
 }
